+++ core/vm/actor/builtin/states/miner/miner_actor_state.hpp
@@ -16,7 +16,6 @@
 #include "vm/actor/builtin/types/miner/deadline_info.hpp"
 #include "vm/actor/builtin/types/miner/deadlines.hpp"
 #include "vm/actor/builtin/types/miner/miner_info.hpp"
-#include "vm/actor/builtin/types/miner/types.hpp"
 #include "vm/actor/builtin/types/miner/vesting.hpp"
 #include "vm/actor/builtin/types/type_manager/universal.hpp"
