+++ core/vm/actor/builtin/types/miner/deadline.hpp
@@ -15,7 +15,7 @@
 #include "vm/actor/builtin/types/miner/post_result.hpp"
 #include "vm/actor/builtin/types/miner/sectors.hpp"
 #include "vm/actor/builtin/types/miner/termination.hpp"
-#include "vm/actor/builtin/types/miner/types.hpp"
+#include "vm/actor/builtin/types/miner/windowed_post.hpp"
 #include "vm/actor/builtin/types/type_manager/universal.hpp"
 
 namespace fc::vm::actor::builtin::types::miner {
