+++ core/vm/actor/builtin/types/miner/power_pair.hpp
@@ -5,19 +5,11 @@
 
 #pragma once
 
-#include "adt/array.hpp"
-#include "primitives/address/address_codec.hpp"
-#include "primitives/rle_bitset/rle_bitset.hpp"
-#include "primitives/sector/sector.hpp"
+#include "codec/cbor/streams_annotation.hpp"
 #include "primitives/types.hpp"
 
 namespace fc::vm::actor::builtin::types::miner {
-  using primitives::ChainEpoch;
-  using primitives::RleBitset;
   using primitives::StoragePower;
-  using primitives::TokenAmount;
-  using primitives::address::Address;
-  using primitives::sector::PoStProof;
 
   struct PowerPair {
     StoragePower raw{};
@@ -55,6 +47,10 @@ namespace fc::vm::actor::builtin::types::miner {
       return (raw == other.raw) && (qa == other.qa);
     }
 
+    inline bool operator!=(const PowerPair &other) const {
+      return !(*this == other);
+    }
+
     inline bool isZero() const {
       return (raw == 0) && (qa == 0);
     }
@@ -66,45 +62,6 @@ namespace fc::vm::actor::builtin::types::miner {
       return result;
     }
   };
-  FC_OPERATOR_NOT_EQUAL(PowerPair)
   CBOR_TUPLE(PowerPair, raw, qa)
 
-  struct WorkerKeyChange {
-    /// Must be an ID address
-    Address new_worker;
-    ChainEpoch effective_at{};
-
-    inline bool operator==(const WorkerKeyChange &other) const {
-      return new_worker == other.new_worker
-             && effective_at == other.effective_at;
-    }
-  };
-  CBOR_TUPLE(WorkerKeyChange, new_worker, effective_at)
-
-  enum class CronEventType {
-    kWorkerKeyChange,
-    kProvingDeadline,
-    kProcessEarlyTerminations,
-  };
-
-  struct CronEventPayload {
-    CronEventType event_type;
-  };
-  CBOR_TUPLE(CronEventPayload, event_type)
-
-  struct WindowedPoSt {
-    /**
-     * Partitions proved by this WindowedPoSt.
-     */
-    RleBitset partitions;
-
-    /**
-     * Array of proofs, one per distinct registered proof type present in the
-     * sectors being proven. In the usual case of a single proof type, this
-     * array will always have a single element (independent of number of
-     * partitions).
-     */
-    std::vector<PoStProof> proofs;
-  };
-  CBOR_TUPLE(WindowedPoSt, partitions, proofs)
 }  // namespace fc::vm::actor::builtin::types::miner
